Use JSON output for rust and rust-cargo checkers
This commit replaces the error parsers for the rust and rust-cargo
checkers to use the JSON output instead.

As a side-effect, we skip the "help" messages which promote the use of
`rustc --explain` in the tests.  I'm assuming this is not critical,
since we also always set the error code now.

Closes GH-1035.